feat: add new `nextRetry` time field on the invoice
Update the invoice table and object to include a `nextRetry` property. This
will be used to track when to retry failed invoices. This field is nullable
because it is only set in certain situations.
Add new dependency on joda time across the entire project to allow the
`DateTime` type to be used across the project. This is the data type used
for datetime columns in the Exposed library.