test: add new test for initial data setup
Add a new test to ensure that the initial data gets correctly setup when
using a sqlite database. This test helps detect any errors that might be
introduced into database read/write code.
The code actually connecting to the database and creating the `AntaeusDal`
object was split out into another function to allow it to be used in test
functions.