feat: setup scheduled job to pay invoices
Create a new scheduled job to handle paying all pending and ready invoices
using the billing service. The job delegates the processing to a batch
worker class that can be easily re-used and tested.
Update the main app class and the job setup functions to take the necessary
parameters to setup the job correctly.
Update plan and time tracker.