feat: new invoice validation job
Add a new recurring job to validate invoices before processing.
Add new validation worker to do the bulk of the work when validating
invoices. This worker handles reading invoices from the database, sending
them to the validation service, and saving them back. This allows the same
logic to be re-used in different contexts.
Refactor the existing scheduled jobs to both extend a common base class to
avoid duplication of boilerplate code (mainly saving and reading the worker
from the job execution context). Add a worker interface for both batch
workers for the same reason.
Add tests for the new worker and execution schedule.
Update implementation plan and time tracker.