refactor: improve invoice query function
Refactor the invoice query function in the `AntaeusDal` class to take
several optional parameters and build up a query based on the ones that
were provided. This allows the same function to be used for several
purposes instead of writing several similar query functions.
Update the `InvoiceService` class to use the new query function.
Add a new function to the `InvoiceService` to find all failed invoices
where the retry payment time is before a provided time. This query will be
used by the job retrying failed payments.