test: add more tests around the invoice queries
Add more tests around the invoice query functions in the `AntaeusDal` class
to ensure they work as expected. New tests ensure that when searching with
a `maxRetryPaymentTime`, only invoices where the corresponding field is not
null and before the provided time are returned.
Pull out duplicate code into separate functions.