feat: job to retry paying failed invoices
Add a new recurring job to attempt paying invoices that have failed. The
job queries the database for all invoices that have a `FAILED` status and
a non-null retry payment date that is < "now". Invoices that have failed
and do not have a retry time set are not considered.
Tweak the invoice service fetch function to be more flexible and support
taking a list of status values to query for.
Pull out code handing the payment of each invoice to a shared code file to
allow it to be re-used across both batch workers (payment and retry).
Update implementation plan and time tracker.