refactor: add function to check if invoice is due
Move the function calculating the start of the current month to the `core`
project to be used by the `isInvoiceDue` function on the invoice service.
Refactor the billing and invoice validation services to take an invoice
service as a constructor parameter instead of as a parameter to their
`...AndSave` functions. This allows the billing service to validate that
the invoice is actually due before trying to pay it.
Update all affected code to provide an invoice service when constructing
these classes.