fix: date conversion to json
Update the project to use the latest version of Kotlin and Javalin. This is
required to use the `jsonMapper` function on the Javalin config object.
This allows a custom mapper to be used when reading/writing data to JSON
for the REST APIs. Create a new custom mapper that registers `JodaModule`
instance. This automatically handles the parsing of dates for JodaTime.
Update implementation plan. This way dates are converted to a unix times
value, instead of a JSON representation of the DateTime object.