feat: protect against paying invoices twice in case of error
Refactor the billing service to protect itself against paying the same
invoice twice. This can happen if the invoice gets sent to the payment
provider and is paid, but the updated invoice then fails to save back to
the database. This would leave the invoice in a READY/PENDING state and
would be picked up for payment again.
The solution is to set the invoice to a PROCESSING state before the invoice
is sent to the payment provider. That way it will not get picked up again
for payment if saving the update fails.
Update the recurring job to retry failed payments to detect stuck invoices
and set them to FAILED.
Add tests to ensure this system works as expected.
Refactor some billing system tests to better detect issues.